terraform/modules/apidata/requirements.txt
flask==2.2.5
gunicorn==20.1.0
requests
google-cloud-bigquery
